rewrite Editor tests so they match the updated model (needed for a long time)
